# backend/tests/api/fixtures/openapi_v1.snapshot.json
{
  "components": {
    "schemas": {
...
      "AreaCountResponse": {
        "description": "Count of areas response schema.",
        "properties": {
          "count": {
            "description": "Total number of areas in context of the current SDEP/member state",
            "examples": [
              42
            ],
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "Area.CountResponse",
        "type": "object"
      },
      "AreaListResponse": {
        "description": "List of areas response schema.",
        "properties": {
          "areas": {
            "description": "List of areas in context of the current SDEP/member state",
            "items": {
              "$ref": "#/components/schemas/AreaResponse"
            },
            "title": "Areas",
            "type": "array"
          }
        },
        "required": [
          "areas"
        ],
        "title": "Area.ListResponse",
        "type": "object"
      },
      "AreaResponse": {
        "description": "Area response schema.",
        "properties": {
          "areaId": {
            "description": "Functional ID identifying this area",
            "examples": [
              "959a7439-7cad-4009-96ec-353b44723db9"
            ],
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Areaid",
            "type": "string"
          },
          "areaName": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Display name (optional) of the area",
            "examples": [
              "Amsterdam"
            ],
            "title": "Areaname"
          },
          "competentAuthorityId": {
            "description": "Functional ID of the competent authority that owns the area",
            "examples": [
              "sdep-ca0363"
            ],
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9-]+$",
            "title": "Competentauthorityid",
            "type": "string"
          },
          "competentAuthorityName": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Display name (optional) of the competent authority",
            "examples": [
              "Amsterdam (inclusief Weesp)"
            ],
            "title": "Competentauthorityname"
          },
          "createdAt": {
            "description": "Timestamp when this area version was created (UTC)",
            "examples": [
              "2025-01-01T00:00:00Z"
            ],
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "filename": {
            "description": "Name of the area shapefile (e.g., 'area.zip')",
            "examples": [
              "Amsterdam.zip"
            ],
            "maxLength": 64,
            "title": "Filename",
            "type": "string"
          },
          "regulation": {
            "$ref": "#/components/schemas/Regulation",
            "default": "all",
            "description": "Regulation type of the area - 'listing', 'activity', or 'all'",
            "examples": [
              "all"
            ]
          }
        },
        "required": [
          "areaId",
          "filename",
          "competentAuthorityId",
          "createdAt"
        ],
        "title": "Area.Response",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "Detail of a single error.",
        "properties": {
          "loc": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Location of the error (e.g. ['body', 'fieldName'])",
            "title": "Loc"
          },
          "msg": {
            "description": "Human-readable error message",
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "description": "Error type identifier",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "msg",
          "type"
        ],
        "title": "Error.Detail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Standardized error response format.",
        "properties": {
          "detail": {
            "description": "List of error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Error.Response",
        "type": "object"
      },
      "Regulation": {
        "description": "Regulation type for an area: 'listing', 'activity', or 'all' (covers both).",
        "enum": [
          "listing",
          "activity",
          "all"
        ],
        "title": "Common.Regulation",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "OAuth2ClientCredentials": {
        "flows": {
          "clientCredentials": {
            "scopes": {},
            "tokenUrl": "<token-url>"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "SDEP is a gateway for the electronic transmission of data between online short-term rental platforms (STR) and competent authorities (CA), ensuring timely, reliable and efficient data sharing processes.\n\n- [Github (open source)](https://github.com/SEMICeu/sdep)\n\n- [EU legislation](https://eur-lex.europa.eu/eli/reg/2024/1028/oj/eng)\n\n- [STR Application Profile (STR-AP)](https://semiceu.github.io/STR-AP/releases/1.0.1/)\n\n- [STR origin prototype v0.0.67](https://eu-str.sdep-pilot.eu/swagger/index.html)\n\nContact:\n\n- [boris.dijkmans@rijksoverheid.nl](mailto:boris.dijkmans@rijksoverheid.nl)",
    "title": "Short Term Rental (STR) - Single Digital Entry Point (SDEP)",
    "version": "<frozen-version>"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ca/areas/{areaId}": {
      "delete": {
        "description": "Delete (deactivate) an area by marking it as ended (now, UTC).\n\n**Behavior:**\n- Deletes (deactivates) the area\n- The area will no longer appear in area listings\n- Deleting an already-deleted area returns 404",
        "operationId": "deleteOwnArea",
        "parameters": [
          {
            "in": "path",
            "name": "areaId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9-]+$",
              "title": "Areaid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success - area deleted (deactivate)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "404": {
            "description": "Resource Not Found - area unavailable, already deleted, or not owned by this CA"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error - business rule violation"
          },
          "500": {
            "description": "Internal Server Error - an unexpected issue occurred that prevented the request from being completed"
          },
          "503": {
            "description": "Service Unavailable - temporarily unable to process requests due to overload, maintenance, or dependency issues (database/authorization server)"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentials": []
          }
        ],
        "summary": "Delete (deactivate) an area from the areas collection for the currently authenticated competent authority",
        "tags": [
          "ca"
        ]
      },
      "get": {
        "description": "Get area (shapefile) based on functional ID, scoped to the currently authenticated competent authority.",
        "operationId": "getOwnArea",
        "parameters": [
          {
            "in": "path",
            "name": "areaId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9-]+$",
              "title": "Areaid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {}
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "404": {
            "description": "Resource Not Found - area unavailable, deleted, or not owned by this CA"
          },
          "500": {
            "description": "Internal Server Error - an unexpected issue occurred that prevented the request from being completed"
          },
          "503": {
            "description": "Service Unavailable - temporarily unable to process requests due to overload, maintenance, or dependency issues (database/authorization server)"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentials": []
          }
        ],
        "summary": "Get area (shapefile) for the currently authenticated competent authority",
        "tags": [
          "ca"
        ]
      }
    },
...
    "/str/areas": {
      "get": {
        "description": "Get all areas. By default, returns all areas (unlimited). Use optional pagination parameters to limit results.\n\n**Each area contains:**\n- `areaId`: Functional ID identifying this area\n- `areaName`: Display name (optional) of the area\n- `regulation`: Regulation type of the area - 'listing', 'activity', or 'all'\n- `filename`: Name of the area shapefile (e.g., 'area.zip')\n- `competentAuthorityId`: Functional ID referencing the competent authority that owns the area\n- `competentAuthorityName`: Display name (optional) of the competent authority\n- `createdAt`: Timestamp when this area version was created (UTC)",
        "operationId": "getAreas",
        "parameters": [
          {
            "description": "Number of records to skip (default: 0)",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of records to skip (default: 0)",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Maximum number of records to return (default: unlimited, max: 1000 when specified)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of records to return (default: unlimited, max: 1000 when specified)",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "areas": [
                    {
                      "areaId": "959a7439-7cad-4009-96ec-353b44723db9",
                      "areaName": "Amsterdam",
                      "competentAuthorityId": "sdep-ca0363",
                      "competentAuthorityName": "Amsterdam (inclusief Weesp)",
                      "createdAt": "2025-01-01T00:00:00Z",
                      "filename": "Amsterdam.zip",
                      "regulation": "all"
                    },
                    {
                      "areaId": "904ee15d-70a6-4e69-a704-6018646803a8",
                      "areaName": "Rotterdam",
                      "competentAuthorityId": "sdep-ca0599",
                      "competentAuthorityName": "Rotterdam",
                      "createdAt": "2025-01-01T00:00:00Z",
                      "filename": "Rotterdam.zip",
                      "regulation": "all"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/AreaListResponse"
                }
              }
            },
            "description": "List of areas"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request - invalid query parameters"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "500": {
            "description": "Internal Server Error - an unexpected issue occurred that prevented the request from being completed"
          },
          "503": {
            "description": "Service Unavailable - temporarily unable to process requests due to overload, maintenance, or dependency issues (database/authorization server)"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentials": []
          }
        ],
        "summary": "Get all areas",
        "tags": [
          "str"
        ]
      }
    },
    "/str/areas/count": {
      "get": {
        "description": "Get areas count (optional, to support pagination).",
        "operationId": "countAreas",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaCountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "500": {
            "description": "Internal Server Error - an unexpected issue occurred that prevented the request from being completed"
          },
          "503": {
            "description": "Service Unavailable - temporarily unable to process requests due to overload, maintenance, or dependency issues (database/authorization server)"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentials": []
          }
        ],
        "summary": "Get areas count (optional, to support pagination)",
        "tags": [
          "str"
        ]
      }
    },
    "/str/areas/{areaId}": {
      "get": {
        "description": "Get area (shapefile) based on functional ID.",
        "operationId": "getArea",
        "parameters": [
          {
            "in": "path",
            "name": "areaId",
            "required": true,
            "schema": {
              "title": "Areaid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {}
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized - missing or invalid token"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          },
          "404": {
            "description": "Resource Not Found - area unavailable"
          },
          "500": {
            "description": "Internal Server Error - an unexpected issue occurred that prevented the request from being completed"
          },
          "503": {
            "description": "Service Unavailable - temporarily unable to process requests due to overload, maintenance, or dependency issues (database/authorization server)"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentials": []
          }
        ],
        "summary": "Get area (shapefile)",
        "tags": [
          "str"
        ]
      }
    }
  }
}